sw:drv: fix spec_fpga_exit() validation
If the FPGA bitstream has not been discovered correctly, then
it does not exist a spec_fpga device. This implies that also
the access to the FPGA is forbidden; therefore, the only mechanism to
check if we failed to load the FPGA content is to check if
the spec_fpga pointer is valid.

Signed-off-by: Federico Vaga <federico.vaga@cern.ch>