Reflect current status in CHANGELOG
It is not that useful to update the change log during development.
I put in it information about previous development, before releaseing
the version 1.4 (yes, we start with that number) we have to update this
file.

I choosed to keep it here and visible as a reminder of our duties

Signed-off-by: Federico Vaga <federico.vaga@cern.ch>