sw:drv: handle correctly level interrupts
While drivers are processing a level interrupts, the bit in GPIO_INT_STATUS
get set and it sticks there. Only when drivers processed the level interrupts
we are sure that the bit does not get set again. Therefore, only at this
point we can properly clean the GPIO_INT_STATUS. Then we may have problems,
with the edge interrupts, so we put them (edge interrupts) back in the
GPIO_INT_STATUS. About this last point we are not sure that is going to work,
but in all our installation we are not going to use edge interrupts because
they are risky.

Signed-off-by: Federico Vaga <federico.vaga@cern.ch>