[hdl] fix DDR controller string "TRUE"/"FALSE" settings
The issue comes from the fact that boolean'image() returns lowercase "true"/"false" strings, which
are then not properly matched by the DDR controller core. To fix this, we pull the to_upper()
function from general-cores and use it to convert the output of bolean'image() to uppercase before
passing it to the DDR controller.

Signed-off-by: Dimitris Lampridis <dimitris.lampridis@cern.ch>