sw:drv: fix GN412x HW bug in FCL
Programming cards in parallel does not work quite well. At best the
GN412x chip reports errors on all plugged cards. In other cases it
freezes the PC.

With this patch I am going to serialize the FCL access. Like this
cards will be programmed one after the other.

I do not know the origin of the bug, but it must be in the GN412x chip.

Signed-off-by: Federico Vaga <federico.vaga@cern.ch>