use modern Python packaging
This is not however 100% compatible with acc_py. However, we are not
planning to release it in acc_py any time soon. By the time this will
happen, probably acc_py will upgrade to support also this approach. And
by the way there is also the old setup.py around.

Signed-off-by: Federico Vaga <federico.vaga@cern.ch>