drv: be sure to have a DMA transfer descriptor
It happened at least once that we got an invalid tx pointer. Not sure
why this happened (e.g. spurious interrupt?), but it happened. With this
patch, we do not prevent this from happening, but we avoid to have a
kernel crash.

This could help to understand the real problem.

Signed-off-by: Federico Vaga <federico.vaga@cern.ch>